container-src/k3s-autoscaler/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
autopage==0.5.2
    # via cliff
cachetools==6.2.3
    # via google-auth
certifi==2025.11.12
    # via
    #   kubernetes
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
cliff==4.13.0
    # via
    #   osc-lib
    #   python-openstackclient
cmd2==3.0.0
    # via cliff
cryptography==46.0.3
    # via
    #   openstacksdk
    #   python-openstackclient
debtcollector==3.0.0
    # via
    #   oslo-utils
    #   python-keystoneclient
decorator==5.2.1
    # via
    #   dogpile-cache
    #   openstacksdk
dogpile-cache==1.5.0
    # via openstacksdk
durationpy==0.10
    # via kubernetes
gnureadline==8.2.13
    # via cmd2
google-auth==2.43.0
    # via kubernetes
idna==3.11
    # via requests
iso8601==2.1.0
    # via
    #   keystoneauth1
    #   openstacksdk
    #   oslo-utils
    #   python-openstackclient
jinja2==3.1.6
    # via -r requirements.in
jmespath==1.0.1
    # via openstacksdk
jsonpatch==1.33
    # via openstacksdk
jsonpointer==3.0.0
    # via jsonpatch
keystoneauth1==5.12.0
    # via
    #   openstacksdk
    #   osc-lib
    #   python-cinderclient
    #   python-keystoneclient
kubernetes==34.1.0
    # via
    #   -r requirements.in
    #   openshift
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.1.2
    # via oslo-serialization
netaddr==1.3.0
    # via
    #   oslo-config
    #   oslo-utils
oauthlib==3.3.1
    # via requests-oauthlib
openshift==0.13.2
    # via -r requirements.in
openstacksdk==4.8.0
    # via
    #   osc-lib
    #   python-openstackclient
os-service-types==1.8.2
    # via
    #   keystoneauth1
    #   openstacksdk
osc-lib==4.2.0
    # via python-openstackclient
oslo-config==10.1.0
    # via python-keystoneclient
oslo-i18n==6.7.1
    # via
    #   osc-lib
    #   oslo-config
    #   oslo-utils
    #   python-cinderclient
    #   python-keystoneclient
    #   python-openstackclient
oslo-serialization==5.8.0
    # via python-keystoneclient
oslo-utils==9.2.0
    # via
    #   osc-lib
    #   oslo-serialization
    #   python-cinderclient
    #   python-keystoneclient
packaging==25.0
    # via
    #   oslo-utils
    #   python-keystoneclient
pbr==7.0.3
    # via
    #   keystoneauth1
    #   openstacksdk
    #   os-service-types
    #   osc-lib
    #   oslo-i18n
    #   oslo-utils
    #   python-cinderclient
    #   python-keystoneclient
    #   python-openstackclient
platformdirs==4.5.1
    # via openstacksdk
prettytable==3.17.0
    # via
    #   cliff
    #   python-cinderclient
psutil==7.1.3
    # via
    #   openstacksdk
    #   oslo-utils
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pygments==2.19.2
    # via rich
pyparsing==3.2.5
    # via oslo-utils
pyperclip==1.11.0
    # via cmd2
python-cinderclient==9.8.0
    # via python-openstackclient
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   kubernetes
python-keystoneclient==5.7.0
    # via python-openstackclient
python-openstackclient==8.2.0
    # via -r requirements.in
python-string-utils==1.0.0
    # via openshift
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   cliff
    #   kubernetes
    #   openstacksdk
    #   oslo-config
    #   oslo-utils
requests==2.32.5
    # via
    #   keystoneauth1
    #   kubernetes
    #   osc-lib
    #   oslo-config
    #   python-cinderclient
    #   python-keystoneclient
    #   python-openstackclient
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via kubernetes
requestsexceptions==1.4.0
    # via openstacksdk
rfc3986==2.0.0
    # via oslo-config
rich==14.2.0
    # via
    #   cmd2
    #   rich-argparse
rich-argparse==1.7.2
    # via cmd2
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via
    #   kubernetes
    #   openshift
    #   python-dateutil
stevedore==5.6.0
    # via
    #   cliff
    #   dogpile-cache
    #   keystoneauth1
    #   osc-lib
    #   oslo-config
    #   python-cinderclient
    #   python-keystoneclient
    #   python-openstackclient
typing-extensions==4.15.0
    # via
    #   keystoneauth1
    #   openstacksdk
    #   os-service-types
tzdata==2025.3
    # via
    #   oslo-serialization
    #   oslo-utils
urllib3==2.3.0
    # via
    #   kubernetes
    #   requests
wcwidth==0.2.14
    # via prettytable
websocket-client==1.9.0
    # via kubernetes
wrapt==2.0.1
    # via debtcollector

# The following packages are considered to be unsafe in a requirements file:
# setuptools
